\name{HPdclassifier-package}
\alias{HPdclassifier-package}
\alias{HPdclassifier}
\docType{package}
\title{Distributed algorithms for classifiers}
\description{
\pkg{HPdclassifier} provides several distributed algorithms for classifiers. It is written based on the infrastructure created in HP-Labs for distributed computing in R.
}

\details{
    \tabular{ll}{
        Package: \tab HPdclassifier\cr
        Type: \tab Package\cr
        Version: \tab 1.0.0\cr
        Date: \tab 2015-01-16\cr
    }
 Main Functions:
 \itemize{
  \item {hpdrandomForest:} {It is a distributed function for randomForest}
  \item {predictHPdRF:} {distributed predict method for applying a random forest objects on a darray or a dframe}
 }
}

\author{
HP Vertica Analytics Team <distributedRTeam@external.groups.hp.com>
}
\references{
 \enumerate{
  \item{Using R for Iterative and Incremental Processing. Shivaram Venkataraman, Indrajit Roy, Alvin AuYoung, Rob Schreiber. HotCloud 2012, Boston, USA.}
 }   
}
\keyword{Distributed R}
\keyword{Scalable Machine Learning algorithms}
\keyword{Big Data Analytics}

